Complete rewrite of TVA, TVF and Pitch classes
To improve the quality of the emulated audio and being able to produce more or
less identical output as the real SC-55 synth, the TVA, TVF and Pitch classes
have been rewritten to follow the exact same fixed point calculations as on
the CPU ROM. This has the added benefit of drastically reducing the CPU usage
and exact audio generation being easier to verify against the real synth.

This rewrite currently handles most of the work needed for these audio classes,
but there are still some rough edges that need more work - and a lot of bug
fixing is expected.

Note that even though this rewrite fixes a number of bugs, improves audio and
timing accuracy, and adds support for new Control ROM parameters, in its
current state the audio is still worse than previous EmuSC versions. This is
due to known shortcomings such as TVF filter not being activated yet, reverb
algorithm is bad and creating distortions etc. Also, only the mk1 ROM version
1.21 has been tested during development.

+++ libemusc/src/part.h
@@ -118,7 +118,7 @@ private:
   // Envelopes and LFOs callback for external clients
   std::function<void(const int)> _changeCallback = NULL;
   std::function<void(const float, const float,
-                     const float, const float,
+                     const int, const int,
                      const float, const float)> _envelopeCallback = NULL;
   std::function<void(const int, const int, const int)> _lfoCallback = NULL;
   bool _callbackTrigger;
